// frontend/src/api/generated/models/timerRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable ticket tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */

export interface TimerRead {
  ticket_id: number;
  ticket_identifier: string;
  ticket_team_key: string;
  ticket_number: number;
  started_at: string;
  duration_seconds: number;
  is_paused: boolean;
  paused_at?: string | null;
  created_at: string;
  updated_at: string;
}